% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_thresh_R.R
\name{iter_thresh_R}
\alias{iter_thresh_R}
\title{Iterative thresholding method for the top eigen decomposition of a matrix}
\usage{
iter_thresh_R(S, u0, theta = 0.1, tol = 1e-04)
}
\arguments{
\item{S}{the matrix whose sparse unit eigenvector to be calculated}

\item{u0}{initial unit vector}

\item{theta}{the sparese parameter which is also the truncation threshold}

\item{tol}{the tolerance of error}
}
\value{
a list of length 2, one is the eigenvector, the other is the estimation for the  (absolute) largest eigenvalue
}
\description{
Iterative thresholding method for calculating sparse unit eigenvector corresponding to the eigenvalue with the largest absolute value
}
\examples{
\dontrun{
S<-matrix(runif(100),nrow=10)
u0<-c(rep(0,9),1)
iter_thresh_R(S,u=u0)
}
}
