% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaco_R.R
\name{Jaco_R}
\alias{Jaco_R}
\title{Jacobian method for calculating the eigenvalues and eigenvectors of a symmetric matrix}
\usage{
Jaco_R(A, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{A}{a symmetric matrix}

\item{tol}{the tolerance of error}

\item{x0}{the initial iteration point}
}
\value{
a list of length 2: values a vector of the estimated eigenvalues, vectors a matrix with each column as the corresponding unit eigenvector
}
\description{
Jacobian method is for calculating the eigenvalues and orthonormal eigenvectors for a symmetric matrix
}
\examples{
\dontrun{
S<-matrix(runif(400),nrow=20)
for(i in 1:19){
for(j in (i+1):20) S[i,j]<-S[j,i]
}
Jaco_R(S,tol=.Machine$double.eps^0.25)
}
}
