% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_Metro_C}
\alias{rw_Metro_C}
\title{Random Walk Metropolis sampler using Rcpp}
\usage{
rw_Metro_C(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the normal random increment with a zero mean}

\item{x0}{the initial iteration point}

\item{N}{the designated number of random numbers (including the initial point x0)}
}
\value{
a list of length 3, x the random numbers of size \code{N}, k the number of rejection times, acceptance_rate the acceptance rate of the candidate points
}
\description{
A Random Walk Metropolis sampler using Rcpp
}
\examples{
\dontrun{
laplace<-function(x) return(1/2*exp(-abs(x)))
lapC <- rw_Metro_C(2,25,2000);
plot(1:2000,lapC$x,type='l')
abline(h=c(-3*sqrt(2),3*sqrt(2)))
}
}
