% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_Metro_R.R
\name{rw_Metro_R}
\alias{rw_Metro_R}
\title{Random Walk Metropolis sampler using R}
\usage{
rw_Metro_R(sigma, x0, N, f)
}
\arguments{
\item{sigma}{the standard deviation of the normal random increment with a zero mean}

\item{x0}{the initial iteration point}

\item{N}{the designated number of random numbers (including the initial point x0)}

\item{f}{the density of target distribution}
}
\value{
a list of length 3, x the random numbers of size \code{N}, k the number of rejection times, acceptance_rate the acceptance rate of the candidate points
}
\description{
A Random Walk Metropolis sampler using R
}
\examples{
\dontrun{
laplace<-function(x) return(1/2*exp(-abs(x)))
lapR <- rw_Metro_R(2,25,2000,laplace);
plot(1:2000,lapR$x,type='l')
abline(h=c(-3*sqrt(2),3*sqrt(2)))
}
}
