\name{SC19085-package}
\alias{SC19085-package}
\alias{SC19085}
\docType{package}
\title{
   This package provides several functions for implementing random walk metropolis sampling, Jacobian and iterative thresholding methods, meanwhile for benchmarking R and Rcpp.
}
\description{
   R and Rcpp versions of functions used to implement random walk metropolis sampler, Jacobian and iterative thresholding methods for calculationg the eigenvalues and eigenvectors are provided in this package (\code{rw_Metro_R}, \code{Jaco_R} and \code{iter_thresh_R} versus \code{rw_Metro_C}, \code{Jaco_C} and \code{iter_thresh_C})
}

\author{
Jinming Liu

Maintainer: Jinming Liu <ljm1997@mail.ustc.edu.cn>
}

\keyword{ package }
